Pull backgrounds controller logic into a facade
